+++ eslint-bridge/src/rules/file.js
@@ -1,0 +1,1 @@
+await 42;
